rename package name from prompt_gen to the other one
