+++ auto_prompt_gen/__init__.py
@@ -1,1 +1,3 @@
 from .generator import PromptGenerator
+
+__all__ = ["PromptGenerator"]
