+++ auto_prompt_gen/about.py
@@ -1,4 +1,4 @@
-__title__ = "prompt_generator"
+__title__ = "auto_prompt_generator"
 __version__ = "0.0.3"
 __summary__ = "A lightweight prompt generator for Python. Primarily for text to image synthesis."
 __uri__ = "https://github.com/Just-Understanding-Data-Ltd/promp_gen"
