+++ auto_prompt_gen/character_data/characters.json
