+++ auto_prompt_gen/location_data/backgrounds.json
