+++ auto_prompt_gen/tests/test_generator.py
@@ -1,4 +1,4 @@
-from prompt_gen import PromptGenerator
+from auto_prompt_gen import PromptGenerator
 import pytest
